refactor: Prepare repository for publishing
